+++ pennylane/devices/autograd_ops.py
@@ -180,3 +180,65 @@ def MultiRZ(theta, n):
         array[complex]: diagonal part of the multi-qubit rotation matrix
     """
     return np.exp(-1j * theta / 2 * pauli_eigs(n))
+
+
+def DoubleExcitation(phi):
+    r"""
+    Double excitation rotation.
+    Args:
+        phi (float): rotation angle
+    Returns:
+        array[complex]: Double excitation rotation matrix
+    """
+    c = np.cos(phi / 2)
+    s = np.sin(phi / 2)
+
+    U = np.zeros((16, 16))
+    U[3, 3] = c  # 3 (dec) = 0011 (bin)
+    U[3, 12] = -s  # 12 (dec) = 1100 (bin)
+    U[12, 3] = s
+    U[12, 12] = c
+
+    return U
+
+
+def DoubleExcitationPlus(phi):
+    r"""
+    Double excitation rotation with positive phase-shift.
+    Args:
+        phi (float): rotation angle
+    Returns:
+        array[complex]: rotation matrix
+    """
+    c = np.cos(phi / 2)
+    s = np.sin(phi / 2)
+    e = np.exp(1j * phi /2)
+
+    U = e * np.eye(16)
+    U[3, 3] = c  # 3 (dec) = 0011 (bin)
+    U[3, 12] = -s  # 12 (dec) = 1100 (bin)
+    U[12, 3] = s
+    U[12, 12] = c
+
+    return U
+
+
+def DoubleExcitationMinus(phi):
+    r"""
+    Double excitation rotation with negative phase-shift.
+    Args:
+        phi (float): rotation angle
+    Returns:
+        array[complex]: rotation matrix
+    """
+    c = np.cos(phi / 2)
+    s = np.sin(phi / 2)
+    e = np.exp(-1j * phi /2)
+
+    U = e * np.eye(16)
+    U[3, 3] = c  # 3 (dec) = 0011 (bin)
+    U[3, 12] = -s  # 12 (dec) = 1100 (bin)
+    U[12, 3] = s
+    U[12, 12] = c
+
+    return U