+++ pennylane/devices/default_qubit_jax.py
@@ -145,6 +145,9 @@ class DefaultQubitJax(DefaultQubit):
         "CRY": jax_ops.CRY,
         "CRZ": jax_ops.CRZ,
         "MultiRZ": jax_ops.MultiRZ,
+        "DoubleExcitation": jax_ops.DoubleExcitation,
+        "DoubleExcitationPlus": jax_ops.DoubleExcitationPlus,
+        "DoubleExcitationMinus": jax_ops.DoubleExcitationMinus,
     }
 
     C_DTYPE = jnp.complex64