+++ pennylane/devices/default_qubit_tf.py
@@ -143,6 +143,9 @@ class DefaultQubitTF(DefaultQubit):
         "CRY": tf_ops.CRY,
         "CRZ": tf_ops.CRZ,
         "CRot": tf_ops.CRot,
+        "DoubleExcitation": tf_ops.DoubleExcitation,
+        "DoubleExcitationPlus": tf_ops.DoubleExcitationPlus,
+        "DoubleExcitationMinus": tf_ops.DoubleExcitationMinus,
     }
 
     C_DTYPE = tf.complex128